+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Adds scroll controllers for picked items & showed items.
 - Adds scroll physics for picked items & showed items.
+- Adds a choice whether to sort showed items & picked
 
 ## 1.0.0