+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 - Adds scroll controllers for picked items & showed items.
 - Adds scroll physics for picked items & showed items.
-- Adds a choice whether to sort showed items & picked
+- Adds a choice whether to sort showed items & picked.
+- Adds picked item remove icon parameter.
 
 ## 1.0.0