+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # 1.0.1
 
-- Adds scroll controllers for picked items & showed items.
-- Adds scroll physics for picked items & showed items.
-- Adds a choice whether to sort showed items & picked.
+- Adds scroll controllers for showed & picked items.
+- Adds scroll physics for showed & picked items.
+- Adds a choice whether to sort showed & picked items.
 - Adds picked item remove icon parameter.
 
 ## 1.0.0