+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # 1.0.1
 
-- Adds scroll controllers for showed & picked items.
-- Adds scroll physics for showed & picked items.
+- Adds [ScrollController] for showed & picked items.
+- Adds [ScrollPhysics] for showed & picked items.
 - Adds a choice whether to sort showed & picked items.
 - Adds picked item remove icon parameter.
+- Exposes picked item's [BoxDecoration].
 
 ## 1.0.0