+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Adds a choice whether to sort showed & picked items.
 - Adds picked item remove icon parameter.
 - Exposes picked item's [BoxDecoration].
+- Exposes clear all & select all buttons' animation curves.
 
 ## 1.0.0